+++ notary/notary.go
@@ -238,7 +238,7 @@ func (n *Notary) send(id string, kind commandspb.NodeSignatureKind, signature []
 	nsig := &commandspb.NodeSignature{Id: id, Sig: signature, Kind: kind}
 	// we send a background context here because the actual context is ignore by the commander
 	// which use a timeout of 5 seconds, this API may need to be addressed another day
-	n.cmd.CommandSync(context.Background(), txn.NodeSignatureCommand, nsig, func(err error) {
+	n.cmd.Command(context.Background(), txn.NodeSignatureCommand, nsig, func(err error) {
 		// just a log is enough here, the transaction will be retried
 		// later
 		n.log.Error("could not send the transaction to tendermint", logging.Error(err))